modularize things
