+++ package.json
@@ -21,6 +21,7 @@
   "homepage": "https://github.com/brekk/throwback#readme",
   "dependencies": {
     "html-webpack-plugin": "^2.30.1",
+    "lodash": "^4.17.4",
     "phaser": "^3.0.0-beta.5",
     "ramda": "^0.25.0"
   },
