+++ src/features.js
@@ -1,0 +1,3 @@
+export const FEATURES = Object.freeze({
+  wiggleShot: true
+})
