+++ src/preload.js
@@ -1,0 +1,3 @@
+export function preload() {
+  this.load.image(`spaceship`, `assets/blue_ship.png`)
+}
