+++ src/world.js
@@ -1,0 +1,11 @@
+export const WORLD = {
+  graphics: null,
+  playerImage: null,
+  playerPhysics: null,
+  bullets: null,
+  up: null,
+  down: null,
+  left: null,
+  right: null,
+  space: null
+}
